# Conflicts:
#	com.unity.render-pipelines.universal/Runtime/ScriptableRenderer.cs
#	com.unity.render-pipelines.universal/ShaderLibrary/Lighting.hlsl
#	com.unity.render-pipelines.universal/ShaderLibrary/ShaderVariablesFunctions.hlsl
#	com.unity.render-pipelines.universal/Shaders/Particles/ParticlesInput.hlsl
#	com.unity.render-pipelines.universal/Shaders/Particles/ParticlesSimpleLitForwardPass.hlsl
#	com.unity.render-pipelines.universal/Shaders/Particles/ParticlesSimpleLitGBufferPass.hlsl
#	com.unity.render-pipelines.universal/Shaders/Particles/ParticlesUnlitForwardPass.hlsl
#	com.unity.render-pipelines.universal/Shaders/SimpleLitForwardPass.hlsl
#	com.unity.render-pipelines.universal/Shaders/Terrain/TerrainLitPasses.hlsl

+++ com.unity.render-pipelines.universal/ShaderLibrary/Lighting.hlsl
@@ -29,7 +29,7 @@ half3 LightingLambert(half3 lightColor, half3 lightDir, half3 normal)
 half3 LightingSpecular(half3 lightColor, half3 lightDir, half3 normal, half3 viewDir, half4 specular, half smoothness)
 {
     float3 halfVec = SafeNormalize(float3(lightDir) + float3(viewDir));
-    half NdotH = saturate(dot(normal, halfVec));
+    half NdotH = half(saturate(dot(normal, halfVec)));
     half modifier = pow(NdotH, smoothness);
     half3 specularReflection = specular.rgb * modifier;
     return lightColor * specularReflection;
